refactor: rename useDomainError → getDomainError and align file names
The hook naming was misleading — the function has no React hook semantics
(no useState/useEffect), so the `use` prefix implied incorrect Rules of
Hooks constraints. Rename to `getDomainError`, rename test-fixture module
to `test-fixtures`, and update all call-sites, docs, and tests to match.
Also tighten DomainError base class with an abstract `code` property and
normalise TrialExpiredError's default message to 'Trial has expired'.
